customer placing order with product details and admin auth for CRUD on products

+++ Shopezly/src/main/java/com/shopezly/repository/ProductRepo.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.shopezly.model.Product;
 
 @Repository
-public interface ProductRepo extends JpaRepository<Product, Long> {
+public interface ProductRepo extends JpaRepository<Product, Integer> {
     
 }
